pyproject.toml
[project]
name = "quantstock-pro"
version = "2.0.0"
description = "Advanced Quantitative Stock Analysis API with AI Insights"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]>=0.29.0",
    "starlette==0.27.0",
    "anyio==3.7.1",
    "httpx==0.25.1",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.2.0",
    "yfinance>=1.0.0",
    "pandas>=2.2.0",
    "pandas-ta>=0.3.14b",
    "numpy>=1.26.0",
    "google-genai>=1.0.0",
    "feedparser>=6.0.11",
    "cachetools>=5.3.0",
    "async-lru>=2.0.4",
    "tenacity>=8.2.3",
    "prometheus-fastapi-instrumentator>=6.1.0",
    "sentry-sdk[fastapi]>=1.40.0",
    "python-dateutil>=2.8.2",
    "newsapi-python>=0.2.7",
    "tomlkit>=0.12.0",
    "rich>=13.7.0",
    "redis>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "black>=24.1.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[tool.setuptools]
packages = ["app"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"